refactor: relocate "lib::ids" to "tui"
As Identifiers are only relevant to the TUI, previously they only still existed due to "IdKey" being required in "Msg".
