+++ lib/src/lib.rs
@@ -7,7 +7,6 @@
 #![allow(clippy::missing_errors_doc)]
 
 pub mod config;
-pub mod ids;
 pub mod invidious;
 pub mod new_database;
 pub mod player;
