+++ tui/src/ui/components/config_editor/general.rs
@@ -24,7 +24,6 @@
 use anyhow::Result;
 use termusiclib::config::SharedTuiSettings;
 use termusiclib::config::v2::tui::{Alignment as XywhAlign, keys::Keys};
-use termusiclib::ids::{Id, IdConfigEditor};
 use tui_realm_stdlib::Radio;
 use tuirealm::props::{Alignment, BorderType, Borders, Color, InputType, Style};
 use tuirealm::{
@@ -35,6 +34,7 @@ use tuirealm::{
 
 use crate::CombinedSettings;
 use crate::ui::components::vendored::tui_realm_stdlib_input::Input;
+use crate::ui::ids::{Id, IdConfigEditor};
 use crate::ui::model::{Model, UserEvent};
 use crate::ui::msg::{ConfigEditorMsg, Msg};
