+++ tui/src/ui/components/popups/deleteconfirm.rs
@@ -1,5 +1,4 @@
 use termusiclib::config::{SharedTuiSettings, TuiOverlay};
-use termusiclib::ids::Id;
 use tuirealm::{
     Component, Event, MockComponent, State, StateValue,
     command::{Cmd, CmdResult, Direction, Position},
@@ -9,6 +8,7 @@ use tuirealm::{
 
 use super::{YNConfirm, YNConfirmStyle};
 use crate::ui::components::vendored::tui_realm_stdlib_input::Input;
+use crate::ui::ids::Id;
 use crate::ui::model::{Model, UserEvent};
 use crate::ui::msg::{DeleteConfirmMsg, Msg};
