+++ tui/src/ui/components/popups/podcast.rs
@@ -1,5 +1,4 @@
 use termusiclib::config::{SharedTuiSettings, TuiOverlay};
-use termusiclib::ids::Id;
 use tui_realm_stdlib::Table;
 use tuirealm::{
     Component, Event, MockComponent, State, StateValue,
@@ -11,6 +10,7 @@ use tuirealm::{
 use super::{YNConfirm, YNConfirmStyle};
 use crate::ui::components::popups::DeleteConfirmInputPopup;
 use crate::ui::components::vendored::tui_realm_stdlib_input::Input;
+use crate::ui::ids::Id;
 use crate::ui::model::{Model, UserEvent};
 use crate::ui::msg::{Msg, PCMsg};
