+++ tui/src/ui/components/popups/saveplaylist.rs
@@ -1,6 +1,5 @@
 use anyhow::Result;
 use termusiclib::config::{SharedTuiSettings, TuiOverlay};
-use termusiclib::ids::Id;
 use tuirealm::{
     Component, Event, MockComponent, State, StateValue,
     command::{Cmd, CmdResult, Direction, Position},
@@ -10,6 +9,7 @@ use tuirealm::{
 
 use super::{YNConfirm, YNConfirmStyle};
 use crate::ui::components::vendored::tui_realm_stdlib_input::Input;
+use crate::ui::ids::Id;
 use crate::ui::model::{Model, UserEvent};
 use crate::ui::msg::{Msg, SavePlaylistMsg};
