+++ tui/src/ui/components/progress.rs
@@ -1,8 +1,6 @@
 use std::time::Duration;
 
-use crate::ui::msg::Msg;
 use termusiclib::config::TuiOverlay;
-use termusiclib::ids::Id;
 use termusiclib::track::DurationFmtShort;
 use termusiclib::track::MediaTypesSimple;
 use tui_realm_stdlib::ProgressBar;
@@ -10,7 +8,9 @@ use tuirealm::props::{Alignment, BorderType, Borders, PropPayload, PropValue};
 use tuirealm::{AttrValue, Attribute, Component, Event, MockComponent};
 
 use crate::ui::Model;
+use crate::ui::ids::Id;
 use crate::ui::model::UserEvent;
+use crate::ui::msg::Msg;
 
 #[derive(MockComponent)]
 pub struct Progress {
