+++ tui/src/ui/ids.rs
@@ -194,3 +194,27 @@ pub enum IdKeyOther {
     PodcastRefreshFeed,
     PodcastRefreshAllFeeds,
 }
+
+#[derive(Debug, Eq, PartialEq, Clone, Copy, Hash)]
+pub enum IdKey {
+    Global(IdKeyGlobal),
+    Other(IdKeyOther),
+}
+
+impl From<&IdKey> for IdConfigEditor {
+    fn from(value: &IdKey) -> Self {
+        match *value {
+            IdKey::Global(id_key_global) => IdConfigEditor::KeyGlobal(id_key_global),
+            IdKey::Other(id_key_other) => IdConfigEditor::KeyOther(id_key_other),
+        }
+    }
+}
+
+impl From<IdKey> for IdConfigEditor {
+    fn from(value: IdKey) -> Self {
+        match value {
+            IdKey::Global(id_key_global) => IdConfigEditor::KeyGlobal(id_key_global),
+            IdKey::Other(id_key_other) => IdConfigEditor::KeyOther(id_key_other),
+        }
+    }
+}
