+++ tui/src/ui/model/mod.rs
@@ -9,7 +9,6 @@ use termusiclib::config::v2::server::ScanDepth;
 use termusiclib::config::v2::tui::keys::Keys;
 use termusiclib::config::v2::tui::theme::ThemeWrap;
 use termusiclib::config::{ServerOverlay, SharedServerSettings, SharedTuiSettings};
-use termusiclib::ids::Id;
 use termusiclib::new_database::Database;
 use termusiclib::new_database::track_ops::TrackRead;
 use termusiclib::player::playlist_helpers::PlaylistTrackSource;
@@ -31,6 +30,7 @@ use super::components::TETrack;
 use super::tui_cmd::TuiCmd;
 use crate::CombinedSettings;
 use crate::ui::Application;
+use crate::ui::ids::Id;
 use crate::ui::model::youtube_options::YoutubeOptions;
 use crate::ui::msg::{Msg, SearchCriteria};
 pub use download_tracker::DownloadTracker;
