+++ tui/src/ui/model/update.rs
@@ -2,7 +2,6 @@ use std::path::Path;
 use std::time::Duration;
 
 use anyhow::anyhow;
-use termusiclib::ids::Id;
 use termusiclib::podcast::{PodcastDLResult, PodcastSyncResult};
 use termusiclib::track::MediaTypesSimple;
 use tokio::runtime::Handle;
@@ -10,6 +9,7 @@ use tokio::time::sleep;
 use tuirealm::Update;
 use tuirealm::props::{AttrValue, Attribute};
 
+use crate::ui::ids::Id;
 use crate::ui::model::youtube_options::YTDLMsg;
 use crate::ui::msg::{
     CoverDLResult, DBMsg, DeleteConfirmMsg, ErrorPopupMsg, GSMsg, HelpPopupMsg, LIMsg, LyricMsg,
