+++ tui/src/ui/model/view.rs
@@ -3,7 +3,6 @@ use std::time::{Duration, Instant};
 
 use anyhow::{Result, bail};
 use termusiclib::config::SharedTuiSettings;
-use termusiclib::ids::{Id, IdConfigEditor, IdTagEditor};
 use termusiclib::utils::get_parent_folder;
 use tokio::runtime::Handle;
 use tui_realm_treeview::Tree;
@@ -19,6 +18,7 @@ use crate::ui::components::{
     FeedsList, Footer, GSInputPopup, GSTablePopup, GlobalListener, LabelSpan, Lyric, MusicLibrary,
     Playlist, Progress, Source,
 };
+use crate::ui::ids::{Id, IdConfigEditor, IdTagEditor};
 use crate::ui::model::{Model, TermusicLayout, UserEvent};
 use crate::ui::msg::{DBMsg, Msg, PCMsg};
 use crate::ui::utils::{
